Reorganize /components directory!
