+++ src/utils/mathUtils.ts
